create test/pylink if system supports symbolic links
Don't package test/pylink (a symbolic link to test/pyfile). Gem
packaging creates a dangling symbolic link. Also, symbolic links
might even be supported on the installation system. Create the
link if it is needed when tests are run. If the system doesn't
support symbolic links, then the link won't be created and the
symbolic link tests will be skipped.